refactor: refactor: 회의에서 짚어주신 컴포넌트 구성형태 올바르게 수정
- index.tsx에 컴포넌트 코드를 넣도록 함
- 컴포넌트 모듈을 가져와서 <C.Components /> 형태로 사용하는건 Tree shaking시 문제가 될 수 있음
  따라서 최적화를 위해 모듈을 각각 import할 수 있도록 함
