+++ src/atoms/Logo/index.tsx
