+++ src/atoms/Text/index.tsx
