+++ src/atoms/index.ts
@@ -1,11 +1,11 @@
-import Button from './Button/Button';
-import Emoji from './Emoji/Emoji';
-import Image from './Image/Image';
-import Logo from './Logo/Logo';
-import Text from './Text/Text';
-import TextArea from './TextArea/TextArea';
-import Title from './Title/Title';
+import Button from './Button';
+import Emoji from './Emoji';
+import Image from './Image';
+import Logo from './Logo';
+import Text from './Text';
+import TextArea from './TextArea';
+import Title from './Title';
 
-export default {
+export {
   Button, Emoji, Image, Logo, Text, TextArea, Title,
 };
